+++ decentraland/components/presentational/EthereumSignIn/EthereumSignIn.js
@@ -1,22 +1,18 @@
 import React from "react";
 import { StyleSheet, View } from "react-native";
-import PropTypes from "prop-types";
 import LargeText from "@presentational/LargeText";
 import WalletButton from "@presentational/WalletButton";
 import Colors from "@constants/Colors";
 
-export default function EthereumSignIn(props) {
+export default function EthereumSignIn() {
   return (
     <View style={styles.container}>
       <LargeText>{`Cool. Let's start by hooking this app with your wallet.`}</LargeText>
-      <WalletButton onConnect={props.onConnect} />
+      <WalletButton appName="gnosis-safe-smart-wallet" scheme="gnosis-safe" />
+      <WalletButton appName="tasit-wallet" scheme="tasit-wallet" />
     </View>
   );
 }
-
-EthereumSignIn.propTypes = {
-  onConnect: PropTypes.func.isRequired,
-};
 
 const styles = StyleSheet.create({
   container: {
