+++ decentraland/components/presentational/EthereumSignIn/EthereumSignIn.test.js
@@ -4,7 +4,6 @@ import EthereumSignIn from "@presentational/EthereumSignIn";
 
 describe("EthereumSignIn", () => {
   it("renders the component", async () => {
-    const onConnect = () => {};
-    expect(shallow(<EthereumSignIn onConnect={onConnect} />)).toMatchSnapshot();
+    expect(shallow(<EthereumSignIn />)).toMatchSnapshot();
   });
 });
